# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nebula-mgmt API
  version: "1.0.0"
  description: >
    Control-plane REST API consumed by the CLI, the nebula-agent poller, and
    mobile bundles. Agent endpoints (enroll, agent/updates) are unauthenticated
    at the bearer layer — enroll carries a single-use token, polls carry an
    Ed25519 proof-of-possession signature (ADR 0004). Admin endpoints require an
    operator API key.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /api/v1/enroll:
    post:
      operationId: enroll
      summary: Enroll a pending host and receive its signed certificate + config.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "200":
          description: Enrollment succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/agent/updates:
    get:
      operationId: agentUpdates
      summary: Signed poll for pending certificate / config / revocation updates.
      description: >
        Requires the four ADR 0004 proof-of-possession headers
        (X-Nebula-Fingerprint, X-Nebula-Timestamp, X-Nebula-Nonce,
        X-Nebula-Signature) rather than a bearer token.
      security: []
      parameters:
        - { in: header, name: X-Nebula-Fingerprint, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Timestamp, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Nonce, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Signature, required: true, schema: { type: string } }
      responses:
        "200":
          description: Current update state for the polling host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentUpdatesResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/agent/config-ack/{version}:
    post:
      operationId: acknowledgeAgentConfig
      summary: Acknowledge a locally validated and written config version.
      description: Uses the same signed proof-of-possession headers as agent polling.
      security: []
      parameters:
        - { in: path, name: version, required: true, schema: { type: integer, minimum: 1 } }
        - { in: header, name: X-Nebula-Fingerprint, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Timestamp, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Nonce, required: true, schema: { type: string } }
        - { in: header, name: X-Nebula-Signature, required: true, schema: { type: string } }
      responses:
        "200":
          description: Version applied or already acknowledged.
          content:
            application/json:
              schema:
                type: object
                required: [config_version]
                properties:
                  config_version: { type: integer, minimum: 1 }
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/agent/import/challenge:
    post:
      operationId: createAgentImportChallenge
      summary: Validate an existing Nebula host snapshot and create a proof-of-possession challenge.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentImportPayload"
      responses:
        "201":
          description: Short-lived host-key challenge created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentImportChallengeResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/agent/import:
    post:
      operationId: registerAgentImport
      summary: Register a validated existing Nebula host in a collecting import session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentImportRequest"
      responses:
        "200":
          description: Idempotent retry returned the existing imported host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentImportResponse"
        "201":
          description: Existing host registered in importing state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentImportResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/networks:
    get:
      operationId: listNetworks
      summary: List networks visible to the operator.
      responses:
        "200":
          description: Networks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Network"
    post:
      operationId: createNetwork
      summary: Create a network.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNetworkRequest"
      responses:
        "201":
          description: Created network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/networks/{id}:
    get:
      operationId: getNetwork
      summary: Get a network by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/hosts:
    get:
      operationId: listHosts
      summary: List hosts visible to the operator.
      responses:
        "200":
          description: Hosts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Host"
    post:
      operationId: createHost
      summary: Create a host and mint its enrollment token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostRequest"
      responses:
        "201":
          description: Created host with one-shot enrollment token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateHostResponse"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/hosts/{id}:
    get:
      operationId: getHost
      summary: Get a host by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/webhook-subscriptions:
    get:
      operationId: listWebhookSubscriptions
      summary: List webhook subscriptions visible to the operator.
      responses:
        "200":
          description: Subscriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookSubscription"
    post:
      operationId: createWebhookSubscription
      summary: Create a webhook subscription.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionRequest"
      responses:
        "201":
          description: Created subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/webhook-subscriptions/{id}:
    get:
      operationId: getWebhookSubscription
      summary: Get a webhook subscription by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateWebhookSubscription
      summary: Update a webhook subscription.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionRequest"
      responses:
        "200":
          description: Updated subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteWebhookSubscription
      summary: Delete a webhook subscription.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/cas/import:
    post:
      operationId: importCA
      summary: Import an existing Nebula CA certificate and signing key.
      description: >
        This secret-bearing endpoint requires direct TLS, literal-loopback
        HTTP, or an explicitly trusted local TLS-terminating proxy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CAImportRequest"
      responses:
        "201":
          description: Existing CA imported.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CA"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "426":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports:
    get:
      operationId: listMeshImports
      summary: List existing-mesh import sessions visible to the operator.
      responses:
        "200":
          description: Import sessions. Bootstrap token hashes are never returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeshImport"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      operationId: createMeshImport
      summary: Start collecting an existing Nebula mesh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeshImportRequest"
      responses:
        "201":
          description: Session created. The reusable bootstrap token is returned only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportTokenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}:
    get:
      operationId: getMeshImport
      summary: Get an existing-mesh import session without its bootstrap token.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Import session with a deterministic reconciliation preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/rotate-token:
    post:
      operationId: rotateMeshImportToken
      summary: Replace a collecting session's reusable bootstrap token.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "201":
          description: Token rotated. The new token is returned only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportTokenResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/cancel:
    post:
      operationId: cancelMeshImport
      summary: Cancel a collecting import and remove only its temporary hosts.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Canceled import session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImport"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/finalize:
    post:
      operationId: finalizeMeshImport
      summary: Atomically accept the current reconciliation preview and begin managed operation.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeshImportFinalizeRequest"
      responses:
        "200":
          description: Finalized import session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImport"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

# Outbound webhooks the server POSTs to an operator-configured endpoint (#256).
# Each delivery carries X-Nebula-Event (type), X-Nebula-Delivery (id), and
# X-Nebula-Signature (sha256=<hmac> over the raw body). The handler/scanner emit
# shapes are validated against these schemas by the contract tests.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Operator API key as a bearer token.
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EnrollRequest:
      type: object
      required: [token, public_key_pem, signing_public_key_pem]
      properties:
        token:
          type: string
          description: Single-use enrollment token. New tokens start with nme_; legacy unprefixed tokens remain valid until expiry.
        public_key_pem:
          type: string
        signing_public_key_pem:
          type: string
        profile:
          $ref: "#/components/schemas/AgentImportProfile"
          description: Optional absolute Nebula file paths and agent capabilities. Omission preserves the /etc/nebula defaults.
    EnrollResponse:
      type: object
      required: [certificate_pem, ca_certificate_pem, config_yaml]
      properties:
        certificate_pem:
          type: string
        ca_certificate_pem:
          type: string
        config_yaml:
          type: string
        config_version:
          type: integer
          minimum: 1
    AgentUpdatesResponse:
      type: object
      required: [has_updates]
      properties:
        has_updates:
          type: boolean
        certificate_pem:
          type: string
        ca_certificate_pem:
          type: string
        config_yaml:
          type: string
        config_version:
          type: integer
          minimum: 1
        blocklist:
          type: array
          items:
            type: string
        import_pending:
          type: boolean
        rekey_required:
          type: boolean
        enrollment_token:
          type: string
...
    AgentImportPayload:
      type: object
      additionalProperties: false
      required: [token, ca_certificate_pem, agent_signing_public_key_pem, payload_hash, snapshot]
      properties:
        token:
          type: string
          pattern: ^nmi_[A-Za-z0-9_-]+$
        ca_certificate_pem: { type: string }
        agent_signing_public_key_pem: { type: string }
        payload_hash:
          type: string
          pattern: ^[a-fA-F0-9]{64}$
        snapshot: { $ref: "#/components/schemas/AgentImportSnapshot" }
    AgentImportRequest:
      type: object
      additionalProperties: false
      required: [token, ca_certificate_pem, agent_signing_public_key_pem, payload_hash, snapshot, challenge_id, proof]
      properties:
        token:
          type: string
          pattern: ^nmi_[A-Za-z0-9_-]+$
        ca_certificate_pem: { type: string }
        agent_signing_public_key_pem: { type: string }
        payload_hash:
          type: string
          pattern: ^[a-fA-F0-9]{64}$
        snapshot: { $ref: "#/components/schemas/AgentImportSnapshot" }
        challenge_id: { type: string }
        proof: { type: string }
    AgentImportChallengeResponse:
      type: object
      required: [challenge_id, session_id, server_public_key, server_nonce, expires_at]
      properties:
        challenge_id: { type: string }
        session_id: { type: string }
        server_public_key: { type: string }
        server_nonce: { type: string }
        expires_at:
          type: string
          format: date-time
    AgentImportResponse:
      type: object
      required: [host_id, certificate_fingerprint, status, created]
      properties:
        host_id: { type: string }
        certificate_fingerprint: { type: string }
        status:
          type: string
          enum: [importing]
        created: { type: boolean }
    CreateNetworkRequest:
      type: object
      required: [name, cidrs]
      properties:
        name:
          type: string
        cidrs:
          type: array
          items:
            type: string
        ca_id:
          type: string
    Network:
      type: object
      required: [id, name, cidrs, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        cidrs:
          type: array
          items:
            type: string
        ca_id:
          type: string
        created_at:
          type: string
          format: date-time
    CreateHostRequest:
      type: object
      required: [network_id, name, nebula_ips]
      properties:
        network_id:
          type: string
        name:
          type: string
        nebula_ips:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        role:
          type: string
        public_ip:
          type: string
        listen_port:
          type: integer
    CreateHostResponse:
      type: object
      required: [host, enrollment_token]
      properties:
        host:
          $ref: "#/components/schemas/Host"
        enrollment_token:
          type: string
          pattern: ^nme_[A-Za-z0-9_-]+$
...